Backend/requirements.txt
# Database
psycopg2-binary>=2.9.9

# Flask Framework
flask==2.3.3
flask-cors==4.0.0
flask-session>=0.5.0
werkzeug>=2.3.0

# Environment & Configuration
python-dotenv==1.0.0

# AI & Machine Learning
openai>=1.0.0
openai-whisper==20231117
sentence-transformers>=5.1.1
huggingface_hub>=0.35.1
scikit-learn>=1.4.0
numpy>=1.26.0

# LangChain & CrewAI
langchain>=0.3.0
langchain-openai>=0.2.0
crewai>=0.203.0

# Voice Processing
pyttsx3==2.90
pydub==0.25.1

# Document Processing
python-docx>=1.0.0
PyPDF2>=3.0.0

# Google OAuth
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1